requirements.txt
pandas
numpy
matplotlib
asammdf
python-docx